Fix MSVC compilation: SDL2, includes, Windows headers
MSVC-specific fixes for clean compilation:
- src/main.cpp: add SDL_MAIN_HANDLED to prevent SDL2 main hijack
- rom_loader.hpp: explicit #include <array> (MSVC doesn't transitively include)
- ppu_viewer_panel.hpp/.cpp, gui_application.cpp: NOMINMAX + WIN32_LEAN_AND_MEAN
  before #include <windows.h> / <GL/gl.h>
- nes_palette.cpp: remove #ifdef NES_GUI_ENABLED guard (always compile ImGui color)
- audio_backend.cpp: explicit narrowing casts for int64_t -> int
- disassembler_panel.cpp: sprintf -> sprintf_s

+++ src/audio/audio_backend.cpp
@@ -28,9 +28,9 @@ bool AudioBackend::initialize(int sample_rate, int buffer_size) {
 	// Configure desired audio specification
 	SDL_AudioSpec desired_spec{};
 	desired_spec.freq = sample_rate;
-	desired_spec.format = AUDIO_F32SYS; // 32-bit float, system byte order
-	desired_spec.channels = 2;			// Stereo
-	desired_spec.samples = buffer_size; // Buffer size in sample frames
+	desired_spec.format = AUDIO_F32SYS;						 // 32-bit float, system byte order
+	desired_spec.channels = 2;								 // Stereo
+	desired_spec.samples = static_cast<Uint16>(buffer_size); // Buffer size in sample frames
 	desired_spec.callback = audio_callback;
 	desired_spec.userdata = this;
